add declared at in inferred type
